backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
requests==2.31.0
googlemaps==4.10.0
APScheduler==3.10.4
python-dotenv==1.0.0
cryptography==41.0.7
